# guest-role/WebContent/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "菜单角色调试接口",
    "version" : "v1",
    "title" : "Guest Role",
    "termsOfService" : "http://www.github.com/kongchen/swagger-maven-plugin",
    "contact" : {
      "name" : "Luo Jing",
      "email" : "luojing@zhiweicloud.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "192.168.1.126:8080",
...
  "schemes" : [ "http" ],
  "paths" : {
...
    "/deleteRoles" : {
      "post" : {
        "tags" : [ "角色管理" ],
        "summary" : "角色管理 - 删除",
        "description" : "返回响应结果",
        "operationId" : "deleteRoles",
        "produces" : [ "application/json;charset=utf8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/全局统一接受参数Long"
          }
        }, {
          "name" : "airportCode",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/list" : {
      "get" : {
        "tags" : [ "角色管理" ],
        "summary" : "角色列表 - 分页查询",
        "description" : "返回分页结果",
        "operationId" : "list",
        "produces" : [ "application/json;charset=utf8" ],
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "default" : 1,
          "format" : "int32"
        }, {
          "name" : "rows",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "default" : 10,
          "format" : "int32"
        }, {
          "name" : "name",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "airportCode",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "请求成功",
            "schema" : {
              "$ref" : "#/definitions/SysRole"
            }
          },
          "400" : {
            "description" : "参数错误"
          },
          "405" : {
            "description" : "请求方式不对"
          }
        }
      }
    },
...
    "/saveOrUpdate" : {
      "post" : {
        "tags" : [ "角色管理" ],
        "summary" : "角色管理 - 新增/修改",
        "description" : "返回成功还是失败",
        "operationId" : "save",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json;charset=utf-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "sysRole",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/全局统一接受参数SysRole"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/view" : {
      "get" : {
        "tags" : [ "角色管理" ],
        "summary" : "角色管理- 根据id查询员工 ",
        "description" : "返回合同详情",
        "operationId" : "view",
        "produces" : [ "application/json;charset=utf8" ],
        "parameters" : [ {
          "name" : "roleId",
          "in" : "query",
          "required" : false,
          "type" : "integer",
          "format" : "int64"
        }, {
          "name" : "airportCode",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "SysRole" : {
      "type" : "object",
      "required" : [ "airportCode", "name" ],
      "properties" : {
        "createTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "创建时间"
        },
        "updateTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "修改时间"
        },
        "isDeleted" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "是否删除：默认为0，0：不删除，1：删除"
        },
        "airportCode" : {
          "type" : "string",
          "description" : "机场code"
        },
        "createUser" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "创建人"
        },
        "updateUser" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "修改人"
        },
        "roleId" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "角色id"
        },
        "name" : {
          "type" : "string",
          "description" : "角色名称"
        },
        "description" : {
          "type" : "string",
          "description" : "角色描述"
        },
        "menuIdList" : {
          "type" : "array",
          "description" : "菜单id集合",
          "items" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "description" : "sys_role"
    },
    "全局统一接受参数Long" : {
      "type" : "object",
      "properties" : {
        "data" : {
          "type" : "array",
          "description" : "数据",
          "items" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "description" : "全局统一接受参数"
    },
...
    "全局统一接受参数SysRole" : {
      "type" : "object",
      "properties" : {
        "data" : {
          "type" : "array",
          "description" : "数据",
          "items" : {
            "$ref" : "#/definitions/SysRole"
          }
        }
      },
      "description" : "全局统一接受参数"
    }
  }
}